+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.2.0 - 2025-09-30
+
+### Added
+
+- Support for live photos, which often have the same base name (e.g. `IMG_0369.HEIC` + `IMG_0369.MOV`)
+
+### Changed
+
+- Less unnecessary warning messages
+
 ## 1.1.1 - 2024-05-23
 
 ### Added